requirements.txt
accelerate
annotated-types
anyio
APScheduler
beautifulsoup4
certifi
charset-normalizer
click
distro
dnspython
email_validator
fastapi
fastapi-cli
filelock
fsspec
h11
hf-xet
httpcore
httptools
httpx
huggingface-hub
idna
IMAPClient
itsdangerous
Jinja2
jiter
jsonpatch
jsonpointer
langchain
langchain-core
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
mdurl
mpmath
networkx
notion-client
numpy
openai
orjson
packaging
psutil
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
python-dotenv
python-multipart
PyYAML
regex
requests
requests-toolbelt
rich
rich-toolkit
safetensors
setuptools
shellingham
sniffio
soupsieve
SQLAlchemy
starlette
sympy
tenacity
tokenizers
torch
tqdm
transformers
typer
typing-inspection
typing_extensions
tzlocal
ujson
urllib3
uvicorn
uvloop
watchfiles
websockets
zstandard
beautifulsoup4
html5lib
python-dotenv